requirements.txt
crewai
langchain-groq
python-dotenv
litellm